When calling GetReaders->SCardListReaders the count of devices that come back is different than calling SCardGetStatusChange.
This is because SCardGetStatusChange only counts readers with an SID that matches the user's SID.

When this happens SCardGetStatusChange  immediately returns causing a CPU spike on the PC.

To solve this we pass in an invalid count of readers calling SCardGetStatusChange  to get an initial count.  Then we call SCardGetStatusChange  again to wait to avoid the CPU spike.